Reduce NVCC threads
